requirements.txt
fastapi
uvicorn
python-jose
pydantic
streamlit
langchain
langchain-community
faiss-cpu
openai
sentence-transformers
pymupdf
python-dotenv
requests
